Fix rapid play bug and add danger zone indicators
If two local players rapidly click on two valid spots, the second play
happens before the first player's animation is complete.  This could
lead to one player accidentally playing for both players.

Now we wait for the animation to complete before accepting click
events on a square.

The "danger zone" indicators mark the corners of the inner 4x4 board.