Fix infinite loop of passing
When neither player can play, the game is over.  But previous to this
change, this implementation would loop forever when that happened,
each player passing forever.

This bug was discovered by my son, and the fix was verified with the
following real-life game transcript I dug up:
  http://transcripts.worldothello.org/woc_history/2017/transcripts/round_05/html/VECCHI%20Maria%20-NICOLAS%20Carol.htm

Running through that transcript, I also realized that I had the first
player backwards.  Black always plays first.  This is now also fixed.