Add novideo mode and connection error handling
Adding ?novideo to the URL will disable audio/video features and let
you play with P2P data only.  This is useful for embedding into other
applications that already host a video chat.

It would still be nice to have a table ID or something similar to
avoid the need to exchange IDs.  Even better would be if that would
permit third and subsequent people to merely watch the game and pick
up the current table state when they "join".