Add tableid support as URL parameter
If there is a table ID in the URL, use that to auto-generate IDs for
the peers and autoconnect them.  This allows joining a game by URL
alone.  Try to make up a unique table ID.

We still need support for game observers, so that additional people
can show up mid-game, get the current game state, and watch the game
in progress.